Move generate stats cloud function sources
