+++ src/cloud/generate-stats/index.js
